@@ -17,7 +17,6 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
@@ -55,13 +54,15 @@ func emitSDK(language, outdir, schemaPath string) error {
 	var generator func() (map[string][]byte, error)
 	switch language {
 	case "dotnet":
-		generator = func() (map[string][]byte, error) { return dotnetgen.GeneratePackage(tool, pkg, extraFiles) }
+		generator = func() (map[string][]byte, error) { return dotnetgen.GeneratePackage(tool, pkg, extraFiles, nil) }
 	case "go":
-		generator = func() (map[string][]byte, error) { return gogen.GeneratePackage(tool, pkg) }
+		generator = func() (map[string][]byte, error) { return gogen.GeneratePackage(tool, pkg, nil) }
 	case "nodejs":
-		generator = func() (map[string][]byte, error) { return nodejsgen.GeneratePackage(tool, pkg, extraFiles) }
+		generator = func() (map[string][]byte, error) {
+			return nodejsgen.GeneratePackage(tool, pkg, extraFiles, nil, false, nil)
+		}
 	case "python":
-		generator = func() (map[string][]byte, error) { return pygen.GeneratePackage(tool, pkg, extraFiles) }
+		generator = func() (map[string][]byte, error) { return pygen.GeneratePackage(tool, pkg, extraFiles, nil) }
 	default:
 		return errors.Errorf("Unrecognized language %q", language)
 	}
@@ -81,7 +82,7 @@ func emitSDK(language, outdir, schemaPath string) error {
 }
 
 func readSchema(schemaPath string) (*schema.Package, error) {
-	schemaBytes, err := ioutil.ReadFile(schemaPath)
+	schemaBytes, err := os.ReadFile(schemaPath)
 	if err != nil {
 		return nil, errors.Wrap(err, "reading schema")
 	}
@@ -91,7 +92,7 @@ func readSchema(schemaPath string) (*schema.Package, error) {
 		return nil, errors.Wrap(err, "unmarshalling schema")
 	}
 
-	pkg, err := schema.ImportSpec(spec, nil)
+	pkg, err := schema.ImportSpec(spec, nil, schema.ValidationOptions{})
 	if err != nil {
 		return nil, errors.Wrap(err, "importing schema")
 	}
@@ -103,7 +104,7 @@ func emitFile(rootDir, filename string, contents []byte) error {
 	if err := os.MkdirAll(filepath.Dir(outPath), 0755); err != nil {
 		return err
 	}
-	if err := ioutil.WriteFile(outPath, contents, 0600); err != nil {
+	if err := os.WriteFile(outPath, contents, 0600); err != nil {
 		return err
 	}
 	return nil
