@@ -2,11 +2,16 @@
 # *** WARNING: this file was generated by Pulumi SDK Generator. ***
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
-import copy
+import builtins as _builtins
 import warnings
+import sys
 import pulumi
 import pulumi.runtime
 from typing import Any, Mapping, Optional, Sequence, Union, overload
+if sys.version_info >= (3, 11):
+    from typing import NotRequired, TypedDict, TypeAlias
+else:
+    from typing_extensions import NotRequired, TypedDict, TypeAlias
 from . import _utilities
 
 __all__ = ['AzureBlobFolderArgs', 'AzureBlobFolder']
@@ -14,22 +19,22 @@ __all__ = ['AzureBlobFolderArgs', 'AzureBlobFolder']
 @pulumi.input_type
 class AzureBlobFolderArgs:
     def __init__(__self__, *,
-                 container_name: pulumi.Input[str],
-                 path: pulumi.Input[str],
-                 resource_group_name: pulumi.Input[str],
-                 storage_account_name: pulumi.Input[str],
-                 disable_managed_object_aliases: Optional[pulumi.Input[bool]] = None,
-                 include_hidden_files: Optional[pulumi.Input[bool]] = None,
-                 managed_objects: Optional[pulumi.Input[bool]] = None):
+                 container_name: pulumi.Input[_builtins.str],
+                 path: pulumi.Input[_builtins.str],
+                 resource_group_name: pulumi.Input[_builtins.str],
+                 storage_account_name: pulumi.Input[_builtins.str],
+                 disable_managed_object_aliases: Optional[pulumi.Input[_builtins.bool]] = None,
+                 include_hidden_files: Optional[pulumi.Input[_builtins.bool]] = None,
+                 managed_objects: Optional[pulumi.Input[_builtins.bool]] = None):
         """
         The set of arguments for constructing a AzureBlobFolder resource.
-        :param pulumi.Input[str] container_name: The name of the Azure storage container to sync to. Required.
-        :param pulumi.Input[str] path: The path (relative or fully-qualified) to the folder containing the files to be synced. Required.
-        :param pulumi.Input[str] resource_group_name: The name of the Azure resource group that the storage account belongs to. Required.
-        :param pulumi.Input[str] storage_account_name: The name of the Azure storage account that the container belongs to. Required.
-        :param pulumi.Input[bool] disable_managed_object_aliases: Disables adding an [alias](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/) resource option to managed objects in the bucket.
-        :param pulumi.Input[bool] include_hidden_files: Include hidden files ("dotfiles") when synchronizing folders. Defaults to `false`.
-        :param pulumi.Input[bool] managed_objects: Whether to have Pulumi manage files as individual cloud resources. Defaults to `true`.
+        :param pulumi.Input[_builtins.str] container_name: The name of the Azure storage container to sync to. Required.
+        :param pulumi.Input[_builtins.str] path: The path (relative or fully-qualified) to the folder containing the files to be synced. Required.
+        :param pulumi.Input[_builtins.str] resource_group_name: The name of the Azure resource group that the storage account belongs to. Required.
+        :param pulumi.Input[_builtins.str] storage_account_name: The name of the Azure storage account that the container belongs to. Required.
+        :param pulumi.Input[_builtins.bool] disable_managed_object_aliases: Disables adding an [alias](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/) resource option to managed objects in the bucket.
+        :param pulumi.Input[_builtins.bool] include_hidden_files: Include hidden files ("dotfiles") when synchronizing folders. Defaults to `false`.
+        :param pulumi.Input[_builtins.bool] managed_objects: Whether to have Pulumi manage files as individual cloud resources. Defaults to `true`.
         """
         pulumi.set(__self__, "container_name", container_name)
         pulumi.set(__self__, "path", path)
@@ -42,115 +47,116 @@ class AzureBlobFolderArgs:
         if managed_objects is not None:
             pulumi.set(__self__, "managed_objects", managed_objects)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="containerName")
-    def container_name(self) -> pulumi.Input[str]:
+    def container_name(self) -> pulumi.Input[_builtins.str]:
         """
         The name of the Azure storage container to sync to. Required.
         """
         return pulumi.get(self, "container_name")
 
     @container_name.setter
-    def container_name(self, value: pulumi.Input[str]):
+    def container_name(self, value: pulumi.Input[_builtins.str]):
         pulumi.set(self, "container_name", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter
-    def path(self) -> pulumi.Input[str]:
+    def path(self) -> pulumi.Input[_builtins.str]:
         """
         The path (relative or fully-qualified) to the folder containing the files to be synced. Required.
         """
         return pulumi.get(self, "path")
 
     @path.setter
-    def path(self, value: pulumi.Input[str]):
+    def path(self, value: pulumi.Input[_builtins.str]):
         pulumi.set(self, "path", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="resourceGroupName")
-    def resource_group_name(self) -> pulumi.Input[str]:
+    def resource_group_name(self) -> pulumi.Input[_builtins.str]:
         """
         The name of the Azure resource group that the storage account belongs to. Required.
         """
         return pulumi.get(self, "resource_group_name")
 
     @resource_group_name.setter
-    def resource_group_name(self, value: pulumi.Input[str]):
+    def resource_group_name(self, value: pulumi.Input[_builtins.str]):
         pulumi.set(self, "resource_group_name", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="storageAccountName")
-    def storage_account_name(self) -> pulumi.Input[str]:
+    def storage_account_name(self) -> pulumi.Input[_builtins.str]:
         """
         The name of the Azure storage account that the container belongs to. Required.
         """
         return pulumi.get(self, "storage_account_name")
 
     @storage_account_name.setter
-    def storage_account_name(self, value: pulumi.Input[str]):
+    def storage_account_name(self, value: pulumi.Input[_builtins.str]):
         pulumi.set(self, "storage_account_name", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="disableManagedObjectAliases")
-    def disable_managed_object_aliases(self) -> Optional[pulumi.Input[bool]]:
+    def disable_managed_object_aliases(self) -> Optional[pulumi.Input[_builtins.bool]]:
         """
         Disables adding an [alias](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/) resource option to managed objects in the bucket.
         """
         return pulumi.get(self, "disable_managed_object_aliases")
 
     @disable_managed_object_aliases.setter
-    def disable_managed_object_aliases(self, value: Optional[pulumi.Input[bool]]):
+    def disable_managed_object_aliases(self, value: Optional[pulumi.Input[_builtins.bool]]):
         pulumi.set(self, "disable_managed_object_aliases", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="includeHiddenFiles")
-    def include_hidden_files(self) -> Optional[pulumi.Input[bool]]:
+    def include_hidden_files(self) -> Optional[pulumi.Input[_builtins.bool]]:
         """
         Include hidden files ("dotfiles") when synchronizing folders. Defaults to `false`.
         """
         return pulumi.get(self, "include_hidden_files")
 
     @include_hidden_files.setter
-    def include_hidden_files(self, value: Optional[pulumi.Input[bool]]):
+    def include_hidden_files(self, value: Optional[pulumi.Input[_builtins.bool]]):
         pulumi.set(self, "include_hidden_files", value)
 
-    @property
+    @_builtins.property
     @pulumi.getter(name="managedObjects")
-    def managed_objects(self) -> Optional[pulumi.Input[bool]]:
+    def managed_objects(self) -> Optional[pulumi.Input[_builtins.bool]]:
         """
         Whether to have Pulumi manage files as individual cloud resources. Defaults to `true`.
         """
         return pulumi.get(self, "managed_objects")
 
     @managed_objects.setter
-    def managed_objects(self, value: Optional[pulumi.Input[bool]]):
+    def managed_objects(self, value: Optional[pulumi.Input[_builtins.bool]]):
         pulumi.set(self, "managed_objects", value)
 
 
+@pulumi.type_token("synced-folder:index:AzureBlobFolder")
 class AzureBlobFolder(pulumi.ComponentResource):
     @overload
     def __init__(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 container_name: Optional[pulumi.Input[str]] = None,
-                 disable_managed_object_aliases: Optional[pulumi.Input[bool]] = None,
-                 include_hidden_files: Optional[pulumi.Input[bool]] = None,
-                 managed_objects: Optional[pulumi.Input[bool]] = None,
-                 path: Optional[pulumi.Input[str]] = None,
-                 resource_group_name: Optional[pulumi.Input[str]] = None,
-                 storage_account_name: Optional[pulumi.Input[str]] = None,
+                 container_name: Optional[pulumi.Input[_builtins.str]] = None,
+                 disable_managed_object_aliases: Optional[pulumi.Input[_builtins.bool]] = None,
+                 include_hidden_files: Optional[pulumi.Input[_builtins.bool]] = None,
+                 managed_objects: Optional[pulumi.Input[_builtins.bool]] = None,
+                 path: Optional[pulumi.Input[_builtins.str]] = None,
+                 resource_group_name: Optional[pulumi.Input[_builtins.str]] = None,
+                 storage_account_name: Optional[pulumi.Input[_builtins.str]] = None,
                  __props__=None):
         """
         Create a AzureBlobFolder resource with the given unique name, props, and options.
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[str] container_name: The name of the Azure storage container to sync to. Required.
-        :param pulumi.Input[bool] disable_managed_object_aliases: Disables adding an [alias](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/) resource option to managed objects in the bucket.
-        :param pulumi.Input[bool] include_hidden_files: Include hidden files ("dotfiles") when synchronizing folders. Defaults to `false`.
-        :param pulumi.Input[bool] managed_objects: Whether to have Pulumi manage files as individual cloud resources. Defaults to `true`.
-        :param pulumi.Input[str] path: The path (relative or fully-qualified) to the folder containing the files to be synced. Required.
-        :param pulumi.Input[str] resource_group_name: The name of the Azure resource group that the storage account belongs to. Required.
-        :param pulumi.Input[str] storage_account_name: The name of the Azure storage account that the container belongs to. Required.
+        :param pulumi.Input[_builtins.str] container_name: The name of the Azure storage container to sync to. Required.
+        :param pulumi.Input[_builtins.bool] disable_managed_object_aliases: Disables adding an [alias](https://www.pulumi.com/docs/intro/concepts/resources/options/aliases/) resource option to managed objects in the bucket.
+        :param pulumi.Input[_builtins.bool] include_hidden_files: Include hidden files ("dotfiles") when synchronizing folders. Defaults to `false`.
+        :param pulumi.Input[_builtins.bool] managed_objects: Whether to have Pulumi manage files as individual cloud resources. Defaults to `true`.
+        :param pulumi.Input[_builtins.str] path: The path (relative or fully-qualified) to the folder containing the files to be synced. Required.
+        :param pulumi.Input[_builtins.str] resource_group_name: The name of the Azure resource group that the storage account belongs to. Required.
+        :param pulumi.Input[_builtins.str] storage_account_name: The name of the Azure storage account that the container belongs to. Required.
         """
         ...
     @overload
@@ -175,13 +181,13 @@ class AzureBlobFolder(pulumi.ComponentResource):
     def _internal_init(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 container_name: Optional[pulumi.Input[str]] = None,
-                 disable_managed_object_aliases: Optional[pulumi.Input[bool]] = None,
-                 include_hidden_files: Optional[pulumi.Input[bool]] = None,
-                 managed_objects: Optional[pulumi.Input[bool]] = None,
-                 path: Optional[pulumi.Input[str]] = None,
-                 resource_group_name: Optional[pulumi.Input[str]] = None,
-                 storage_account_name: Optional[pulumi.Input[str]] = None,
+                 container_name: Optional[pulumi.Input[_builtins.str]] = None,
+                 disable_managed_object_aliases: Optional[pulumi.Input[_builtins.bool]] = None,
+                 include_hidden_files: Optional[pulumi.Input[_builtins.bool]] = None,
+                 managed_objects: Optional[pulumi.Input[_builtins.bool]] = None,
+                 path: Optional[pulumi.Input[_builtins.str]] = None,
+                 resource_group_name: Optional[pulumi.Input[_builtins.str]] = None,
+                 storage_account_name: Optional[pulumi.Input[_builtins.str]] = None,
                  __props__=None):
         opts = pulumi.ResourceOptions.merge(_utilities.get_resource_opts_defaults(), opts)
         if not isinstance(opts, pulumi.ResourceOptions):
