@@ -2,11 +2,16 @@
 # *** WARNING: this file was generated by Pulumi SDK Generator. ***
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
-import copy
+import builtins as _builtins
 import warnings
+import sys
 import pulumi
 import pulumi.runtime
 from typing import Any, Mapping, Optional, Sequence, Union, overload
+if sys.version_info >= (3, 11):
+    from typing import NotRequired, TypedDict, TypeAlias
+else:
+    from typing_extensions import NotRequired, TypedDict, TypeAlias
 from . import _utilities
 
 __all__ = ['ProviderArgs', 'Provider']
@@ -20,6 +25,7 @@ class ProviderArgs:
         pass
 
 
+@pulumi.type_token("pulumi:providers:synced-folder")
 class Provider(pulumi.ProviderResource):
     @overload
     def __init__(__self__,
